Rename and update unit tests
